[big-review] - upgrade kotlin and gradle version - remove momopay
