+++ gradle.properties
@@ -1,5 +1,5 @@
-VERSION_NAME=1.3.0
-VERSION_CODE=50
+VERSION_NAME=1.4.0
+VERSION_CODE=60
 org.gradle.jvmargs=-Xmx1536m
 android.useAndroidX=true
 android.enableJetifier=true
