--- lib/src/main/java/co/opensi/kkiapay/jvm_compatibility.kt
@@ -1,9 +1,0 @@
-package co.opensi.kkiapay
-
-/**
- *  @author Shadai ALI ( shadai.ali@opensi.co )
- *  created  at 6/4/18.
- */
-interface KKiapayCallback {
-    fun onResponse(status: STATUS, phone: String, transactionId: String)
-}
