+++ lib/src/main/res/layout/custom_tab_activity.xml
@@ -5,7 +5,7 @@
     android:background="@android:color/transparent">
 
     <WebView
-        android:id="@+id/kkiapay.web.view"
+        android:id="@+id/kkiapayWebView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@android:color/transparent"
