+++ sample/src/main/java/co/opensi/kkiapay_sdk/MyApp.kt
@@ -1,19 +1,18 @@
-package co.opensi.kkiapay_sdk.kotlin
+package co.opensi.kkiapay_sdk
 
 import android.app.Application
 import co.opensi.kkiapay.uikit.Kkiapay
 import co.opensi.kkiapay.uikit.SdkConfig
-import co.opensi.kkiapay_sdk.R
 
 class MyApp : Application() {
     override fun onCreate() {
         super.onCreate()
         //Initialisation de l'API
         Kkiapay.init(applicationContext,
-                "<PUBLIC-API-KEY>",
+                "LprYUAyMpfAjq4z2yTHPiY0b6XktIQ",
             SdkConfig(themeColor = R.color.colorPrimary,
-                imageResource = R.raw.armoiries,
                     /* set enableSandbox = false in case you are using live API Keys */
-                    enableSandbox = true ))
+                    enableSandbox = false )
+        )
     }
 }
