+++ sample/src/main/res/values/themes.xml
@@ -1,0 +1,9 @@
+<resources xmlns:tools="http://schemas.android.com/tools">
+    <!-- Base application theme. -->
+    <style name="Base.Theme.Kkiapay_android_sdk" parent="Theme.Material3.DayNight.NoActionBar">
+        <!-- Customize your light theme here. -->
+        <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+    </style>
+
+    <style name="Theme.Kkiapay_android_sdk" parent="Base.Theme.Kkiapay_android_sdk" />
+</resources>
